Moved statistics package to backend app
