+++ backend/app/statistics/package.json
@@ -9,10 +9,11 @@
         }
     },
     "scripts": {
-        "build": "tsc --build tsconfig.build.json",
+        "build": "tsc --build tsconfig.build.json && yarn -s copy-assets",
         "build:full": "yarn -s clear && yarn -s build",
         "clear": "rm -rf ./dist && rm -f *.tsbuildinfo",
-        "dev": "wait-on ../../shared/statistics/dist/sync.js && concurrently -r 'yarn -s build --watch --preserveWatchOutput' \"wait-on ./dist/index.js && nodemon --quiet --inspect=5859 --watch dist --watch '../../../shared/*/dist/' --watch '../../shared/*/dist/' --ext .ts,.json,.sql,.js --watch .env.json --delay 1000ms --exec 'node --enable-source-maps ./dist/index.js' --signal SIGTERM\"",
+        "copy-assets": "mkdir -p ./dist/migrations ./dist/report && rsync --delete -r --checksum ./migrations/ ./dist/migrations/ && rsync --delete -r --checksum ./report/ ./dist/report/",
+        "dev": "wait-on ../../shared/middleware/dist/index.js && concurrently -r 'yarn -s build --watch --preserveWatchOutput' \"wait-on ./dist/index.js && nodemon --quiet --inspect=5859 --watch dist --watch '../../../shared/*/dist/' --watch '../../shared/*/dist/' --ext .ts,.json,.sql,.js --watch .env.json --delay 1000ms --exec 'node --enable-source-maps ./dist/index.js' --signal SIGTERM\"",
         "dev:build": "yarn -s build",
         "dev:full": "yarn -s dev",
         "lint": "eslint",
@@ -34,7 +35,6 @@
         "@stamhoofd/models": "2.140.0",
         "@stamhoofd/queues": "2.140.0",
         "@stamhoofd/sql": "2.140.0",
-        "@stamhoofd/statistics": "2.140.0",
         "@stamhoofd/structures": "2.140.0",
         "@stamhoofd/types": "2.140.0",
         "@stamhoofd/utility": "2.140.0",
