+++ backend/app/statistics/report/filters.sql
