+++ backend/app/statistics/report/includes/facts-alle-jaren.sql
