+++ backend/app/statistics/report/includes/facts.sql
