+++ backend/app/statistics/report/includes/gtp.sql
