+++ backend/app/statistics/src/boot.ts
@@ -2,8 +2,8 @@ import { Column, Database } from '@simonbackx/simple-database';
 import { I18n } from '@stamhoofd/backend-i18n/I18n';
 import { startCrons, stopCrons, waitForCrons } from '@stamhoofd/crons';
 import { loadLogger } from '@stamhoofd/logging';
-import { endStatisticsConnection } from '@stamhoofd/statistics/connection';
 import { Version } from '@stamhoofd/structures';
+import { endStatisticsConnection } from './connection.js';
 
 process.on('unhandledRejection', (error: Error) => {
     console.error('unhandledRejection');
