+++ backend/app/statistics/src/periods.test.ts
