+++ backend/app/statistics/src/periods.ts
