+++ backend/app/statistics/src/print-report.ts
