+++ backend/app/statistics/src/writer.ts
