--- backend/shared/statistics/.gitignore
@@ -1,221 +1,0 @@
-/devops/
-node_modules/
-.node_modules/
-built/*
-tests/cases/rwc/*
-tests/cases/test262/*
-tests/cases/perf/*
-!tests/cases/webharness/compilerToString.js
-test-args.txt
-~*.docx
-\#*\#
-.\#*
-tests/baselines/local/*
-tests/baselines/local.old/*
-tests/services/baselines/local/*
-tests/baselines/prototyping/local/*
-tests/baselines/rwc/*
-tests/baselines/test262/*
-tests/baselines/reference/projectOutput/*
-tests/baselines/local/projectOutput/*
-tests/baselines/reference/testresults.tap
-tests/services/baselines/prototyping/local/*
-tests/services/browser/typescriptServices.js
-src/harness/*.js
-src/compiler/diagnosticInformationMap.generated.ts
-src/compiler/diagnosticMessages.generated.json
-src/parser/diagnosticInformationMap.generated.ts
-src/parser/diagnosticMessages.generated.json
-rwc-report.html
-*.swp
-build.json
-*.actual
-tests/webTestServer.js
-tests/webTestServer.js.map
-tests/webhost/*.d.ts
-tests/webhost/webtsc.js
-tests/cases/**/*.js
-!tests/cases/docker/*.js/
-tests/cases/**/*.js.map
-*.config
-scripts/eslint/built/
-scripts/debug.bat
-scripts/run.bat
-scripts/word2md.js
-scripts/buildProtocol.js
-scripts/ior.js
-scripts/authors.js
-scripts/configurePrerelease.js
-scripts/configureLanguageServiceBuild.js
-scripts/open-user-pr.js
-scripts/open-cherry-pick-pr.js
-scripts/processDiagnosticMessages.d.ts
-scripts/processDiagnosticMessages.js
-scripts/produceLKG.js
-scripts/importDefinitelyTypedTests/importDefinitelyTypedTests.js
-scripts/generateLocalizedDiagnosticMessages.js
-scripts/request-pr-review.js
-scripts/*.js.map
-scripts/typings/
-coverage/
-internal/
-**/.DS_Store
-.settings
-!tests/cases/projects/projectOption/**/node_modules
-!tests/cases/projects/NodeModulesSearch/**/*
-!tests/baselines/reference/project/nodeModules*/**/*
-.idea
-yarn-error.log
-.parallelperf.*
-tests/cases/user/*/package-lock.json
-tests/cases/user/*/node_modules/
-tests/cases/user/*/**/*.js
-tests/cases/user/*/**/*.js.map
-tests/cases/user/*/**/*.d.ts
-!tests/cases/user/zone.js/
-!tests/cases/user/bignumber.js/
-!tests/cases/user/discord.js/
-tests/baselines/reference/dt
-.failed-tests
-TEST-results.xml
-package-lock.json
-tests/cases/user/TypeScript-React-Starter/TypeScript-React-Starter
-tests/cases/user/TypeScript-Node-Starter/TypeScript-Node-Starter
-tests/cases/user/TypeScript-React-Native-Starter/TypeScript-React-Native-Starter
-tests/cases/user/TypeScript-Vue-Starter/TypeScript-Vue-Starter
-tests/cases/user/TypeScript-WeChat-Starter/TypeScript-WeChat-Starter
-tests/cases/user/create-react-app/create-react-app
-tests/cases/user/fp-ts/fp-ts
-tests/cases/user/webpack/webpack
-tests/cases/user/puppeteer/puppeteer
-tests/cases/user/axios-src/axios-src
-tests/cases/user/prettier/prettier
-.eslintcache
-
-# Logs
-logs
-*.log
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-lerna-debug.log*
-
-# Diagnostic reports (https://nodejs.org/api/report.html)
-report.[0-9]*.[0-9]*.[0-9]*.[0-9]*.json
-
-# Runtime data
-pids
-*.pid
-*.seed
-*.pid.lock
-
-# Directory for instrumented libs generated by jscoverage/JSCover
-lib-cov
-
-# Coverage directory used by tools like istanbul
-coverage
-*.lcov
-
-# nyc test coverage
-.nyc_output
-
-# Grunt intermediate storage (https://gruntjs.com/creating-plugins#storing-task-files)
-.grunt
-
-# Bower dependency directory (https://bower.io/)
-bower_components
-
-# node-waf configuration
-.lock-wscript
-
-# Compiled binary addons (https://nodejs.org/api/addons.html)
-build/Release
-
-# Dependency directories
-node_modules/
-jspm_packages/
-
-# Snowpack dependency directory (https://snowpack.dev/)
-web_modules/
-
-# TypeScript cache
-*.tsbuildinfo
-
-# Optional npm cache directory
-.npm
-
-# Optional eslint cache
-.eslintcache
-
-# Microbundle cache
-.rpt2_cache/
-.rts2_cache_cjs/
-.rts2_cache_es/
-.rts2_cache_umd/
-
-# Optional REPL history
-.node_repl_history
-
-# Output of 'npm pack'
-*.tgz
-
-# Yarn Integrity file
-.yarn-integrity
-
-# dotenv environment variables file
-.env
-.env.test
-
-# parcel-bundler cache (https://parceljs.org/)
-.cache
-.parcel-cache
-
-# Next.js build output
-.next
-
-# Nuxt.js build / generate output
-.nuxt
-dist
-
-# Gatsby files
-.cache/
-# Comment in the public line in if your project uses Gatsby and not Next.js
-# https://nextjs.org/blog/next-9-1#public-directory-support
-# public
-
-# vuepress build output
-.vuepress/dist
-
-# Serverless directories
-.serverless/
-
-# FuseBox cache
-.fusebox/
-
-# DynamoDB Local files
-.dynamodb/
-
-# TernJS port file
-.tern-port
-
-# Stores VSCode versions used for testing VSCode extensions
-.vscode-test
-
-# yarn v2
-
-.yarn/cache
-.yarn/unplugged
-.yarn/build-state.yml
-.pnp.*
-
-.yalc
-yalc.lock
-
-.yarn
-
-# Yarnrc should be removed from git, because it breaks yarn if the version is not installed yet
-.yarnrc
-
-.env.json
-.env.*.json
-!.env.template.json
