--- backend/shared/statistics/tsconfig.build.json
@@ -1,15 +1,0 @@
-{
-    "extends": "../../../tsconfig.base.json",
-    "compilerOptions": {
-        "rootDir": "src",
-        "outDir": "dist"
-    },
-    "include": [
-        "./src",
-        "../../stamhoofd.d.ts"
-    ],
-    "exclude": [
-        "./src/**/*.spec.ts", 
-        "./src/**/*.test.ts"
-    ]
-}
