--- backend/shared/statistics/tsconfig.json
@@ -1,12 +1,0 @@
-{
-    "extends": "../../../tsconfig.base.json",
-    "files": [],
-    "references": [
-        {
-            "path": "./tsconfig.build.json"
-        },
-        {
-            "path": "./tsconfig.test.json"
-        }
-    ]
-}
