--- backend/shared/statistics/tsconfig.test.json
@@ -1,18 +1,0 @@
-{
-    "compilerOptions": {
-        "outDir": "dist",
-        "rootDir": ".",
-        "noEmit": true
-    },
-    "extends": "../../../tsconfig.test.json",
-    "include": [
-        "../../stamhoofd.d.ts",
-        "../../../jest-extended.d.ts",
-        "./src/**/*.spec.ts", 
-        "./src/**/*.test.ts",
-        "./tests"
-    ],
-    "references": [
-        { "path": "./tsconfig.build.json" }
-    ]
-}
